+++ packages/experimental/src/components/PageHeader/ButtonSetWithOverflow.js
@@ -10,12 +10,9 @@ import PropTypes from 'prop-types';
 
 import cx from 'classnames';
 import ReactResizeDetector from 'react-resize-detector';
-
-import { settings } from 'carbon-components';
-const { prefix } = settings;
 import { ButtonSet } from 'carbon-components-react';
 
-import { pkgPrefix } from '../../global/js/settings';
+import { pkgPrefix, carbonPrefix } from '../../global/js/settings';
 // import uuidv4 from '../../global/js/utils/uuidv4';
 // import unwrapIfFragment from '../../global/js/utils/unwrap-if-fragment';
 import { TempComboButton } from './TempComboButton';
@@ -38,7 +35,7 @@ export const ButtonSetWithOverflow = ({
     let newShowAsOverflow = true;
 
     const sizingSet = sizingContainerRefSet.current?.querySelectorAll(
-      `.${prefix}--btn`
+      `.${carbonPrefix}--btn`
     );
 
     let sizingSetTotalSize = 0;